Add useScrollLock hook
This hook moves our scroll lock mechanism to its own module. It also changes how it works because in this setup, it will dynamically figure out the scrollBarCompensation and add it as padding-right to the body element (so when the scrollbar disappears, it is replaced by padding to prevent weird offsets). The previous CSS changes commit also change the background color of the body element so this padding doesn't look weird.

Before this change we always assumed their would be a scrollbar of static size (even if there was no scrollbar at all) so the compensation actually made things worse before. Now it works beautifully