Cleaned up client-side code
The code is much more robust (even though there shouldn't realistically be errors anyways). What I mean by this is when you press 'Submit Survey', it doesn't actually give a fuck if your HTTP request went through. Since the survey results (accuracy percentages) are calculated on the client anyways, it doesn't care if the survey results actually made it to the DB. However by using dev tools and looking at the console, we can see if the request succeeded